pyproject.toml
[build-system]
requires = ["uv_build>=0.7.19,<0.8.0"]
build-backend = "uv_build"

[project]
name = "base-cacheable-class"
version = "0.1.0"
description = "A flexible base class for adding caching capabilities to your Python classes"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
keywords = ["cache", "caching", "decorator", "redis", "in-memory"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "redis>=6.2.0",
    "valkey-glide>=2.0.1",
]

[project.optional-dependencies]
redis = [
    "redis>=5.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/base-cacheable-class"
Documentation = "https://github.com/yourusername/base-cacheable-class#readme"
Repository = "https://github.com/yourusername/base-cacheable-class"
Issues = "https://github.com/yourusername/base-cacheable-class/issues"

[tool.uv.build-backend]
module-root = ""

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "redis.*"
ignore_missing_imports = true

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'

[tool.pyrefly]
python-version = "3.10.0"
replace-imports-with-any = ["redis.*"]

[dependency-groups]
dev = [
    "pytest-mock>=3.14.1",
]
